Fix #1, #2, #5, #9 implements stable version
The FX scripts are added and changed: Now the aiming FX is managed by
the spell and Wld_PlayEffect was removed. (See #5)

The focus vob is no set to the aim vod which is an oCItem. This works
very consistently now. (See #2)

The mouse is not really frame locked but rather scaled to the fps.
(See #1)

The mouse sensitivity issue seems to be resolved by #1. (See #9)