Fix #97 set hit detection to experimental
The addititonal hit detection test is kept in the scripts but is set
to experimental and disabled by default. It performs a manual check
on top of the engine collision detection (which should not be
necessary in the first place) and checks whether the trajectory of
the projectile intersects with the bbox.
However, rather than improving the hit detection, it only caused false
negatives (by my estimation).