Fix #100 critical bug with trace ray behavior
The issue with Blink is now fixed. However, I discovered a strange
peculiarity about the trace ray flags. Although the bug did not
seem to be related to alpha textures, the respective flag was
necessary to fix it.
Unfortunately this fix enables blink to traverse the player through
alpha vobs like cages, fences and bars.