Improve performance, fix spells stopping mid-air
It is now checked whether a trace ray hits an npc directly. In that case
the focus validation is skipped. Previously, trace rays were assumed to
never detect npcs which turns out to be only true in some cases. Skipping
focus validation improves performance as only one trace ray is shot.
Additionally, the bug of spells stopping mid-air has been fixed by setting
the aim vob to the maximum distance (depending on the spell properties)
for spells (ranged combat uneffected). This is possible, because spells
travel in a straight line, unlike projectiles (gravity drop-off). Thanks
to this change, spells will not stop mid-air if the trace ray wrongly
intersects with a too large bbox.