Fix major trace ray bug
The trouble with trace rays and the resulting limitations revolved around
the mistake of defining the trace ray flags as constant. This constant
would be overwritten at a later point compromising the intended effect.
Thus, the trace ray would intersect with bounding boxes along other
problems. This is now fixed, rendering the previous spell fix unnecessary.
This commit marks a major improvement.