Reinstate trace ray alpha poly flag
Unfortunately, the big trace ray flag fix does not resolve the alpha poly
behavior of trace rays. This means the trace ray ignores (static) walls in
some cases. The seemingly unrelated flag "POLY_IGNORE_TRANSP" is necessary
to prevent this behavior. The disadvantage of using this flag is, that the
aim vob ignores intersections with alpha textures, bypassing gates and
fences. This becomes a problem in conjunction with the spell blink
(traverse through gates that are usually inpenetreble). Nevertheless,
g2freeAim by itself is not directly impacted by this and works without any
limitations.