Fix #105 trace ray performance
The cause for the performance drops was an awkwardly hooked function
(see #105) which caused the trace ray machinery to be run ten times per
freame during aiming in ranged combat. This is now resolved (tenfold
speed-up).
Performance can be further boosted by a new entry in the ini file which
introduces an adjustable recalculation frequency in milliseconds.
By default the trace ray/focus collection is recalculated at each frame.
When increasing the interval (to up to 500 ms) the focus collection
becomes less precise, but might boost the performance slightly on weak
machines (powerful machines will not benefit from this setting and
should keep the value close to zero).
Caution: Increasing this value beyond 45 ms will introduce a stutter
in spells like blink that continually visualize the aim vob.