Revert to manual rotation by mouse
The simplified model rotation introduced with d1616a5 is not accurate
enough. The model will only be rotated if the mouse movement was larger
than a certain minimum threshold which makes aiming impossible. Although,
this was a nice commit in terms of performance and cutting down the code,
it is just not sufficient. All changes have been reverted.