Disable auto turning towards enemy
The core functionality of this feature is implemented. However, to make it
work, it needs to be enabled and disabled depending on whether free aiming
is active. This requires reworking freeAimIsActive().
For now, it is done with final() which does not fully work as the stack
seems to be modified, compromising the return value.