Fix #127 missing focus modes and lvl change crash
During level change some functions are called before Init_Global()
causing a crash or at least empty instances if Ikarus is not initialized
before hand. This has been fixed with this commit, along side with
reinitilizing the focus instances (classes of C_Focus as defined in
Focus.d). Strangely enough, Gothic does not bother to reassign the
instances on level change, which is critical for correct focus collection
when switching between free aim and auto aim in the menu.