Rescale scatter with accuracy
The translation from an accuracy percentage to scattering angles is now
more reasonable and linear. This is a huge improvement to the previous
scattering.
Nevertheless, it is not perfect yet, because the loose hit registration
of bounding boxes (instead of the actual models) does not allow 10% bow
talent to correspond to 10% hit chance, while maintaining the ability to
reliably hit critical hits with 100%. Either the scattering percentage is
correct with extremely large scattering, or the hit percentage does not
faithfully represent the accuracy but 100% allows precise shooting.
A compromise would not satisfy. A different solution needs to be found.