Fix gravity drop-off
While reworking the "collectable" feature, the gravity was always
overwritten resulting in extremely fast projectile drop-off. Along with
some small improvements this has been solved in this commit, by hooking
the oCAIArrowBase::DoAI at a certain offset to reset the gravity
independently of the "collectable" function.
This also allows better separatability of the features.