Fix #106 finalize hit chance and accuracy
The scattering is refined and the minimun and maximum scatter angles are
moved to the internal constants, because they should not be adjusted
anymore. When disabling scattering, the entire accuracy machinery is
disabled, because the accuracy value cannot replace the hit chance alone,
since the distance is not considered. Instead, when disabling scattering,
Gothic's default hit chance calculation (based on skill and distance) is
used.