Fix bug with disabled mouse
Since the function freeAimIsActive() has been moved to the mouse movement,
the settings would not be any longer updated when playing with mouse
controls disabled, resulting in broken controls. The menu options
depending settings have now been outsourced into another function that is
only called when the menu options change and on initialization. This fixes
the problem and slightly increases performance as the checks are no longer
performed every frame.