Fix #111 introduce adjustable recoil
Unlike the previous attempt with adding a dynamic cut scene camera for
implementing recoil, it is now achieved by simply manipulating the mouse
movement. This commit adds another step to being better balanced, as by
default (adjustable in the configuration) bows do not have recoil, while
for crossbows it is scaled inversely with strength. This way the
previously superior crossbows (due to always 100% draw force) get nerfed
and the balance between bows and crossbows is reinstated.