Balance critical hits on protected targets
When the proctection of the target is higher than the damage caused by a
projectile, there should never be any damage. However, this was possible
to bypass with critical hits, that increased the damage of the projectile
by a significant factor. This was a balancing issue. This is now fixed, by
not allowing critical hits, when the base damage would not ellicit any
damage (by Gothic's ranged damage formular).