Create proper trail strip PFX
Since Gothic 1 does not have the property C_ParticleFX.useEmittersFOR to
tie to and travel with their emitter, the trail is created by particle
movement against motion (towards the back of the arrow/bolt). This looks
very similar. The downside is the orientation of the arrow while in
setting up the projectile: The FX will be sideways on top of the player
model.
To counter this side effect, no delaying mechanism would work reliably
(e.g. CFx_Base.emFXTriggerDelay and C_ParticleFX.ppsCreateEmDelay). What
eventually did the trick was artifically delaying the PFX with
C_ParticleFX.ppsScaleKeys_S, sacrificing the looping ability (no problem).
In the end, it works nicely, is well visible and resembles what has been
done for Gothic 2.