Fix #141 remove focus when not aiming
With the additional setting "GFA_NO_AIM_NO_FOCUS" it is now possible to
decide to remove the focus when not aiming. Aside from being more
reasonable, this also prevents abusing ranged weapons or spells as enemy
detectors (running through a forest with the bow drawn to pick up on the
nearest threat before actually seeing it).
The default value of the focus update interval (Gothic.ini setting) has
been decreased to instantaneous updates, as the focus started to flicker.
Upon further investigation, this setting might have to be removed if it
will cause flickering.