Fix #170 enlarge humanoid bounding boxes
The function of calculating the bounding box of a model is hijacked and
the bounding box is enlarged if the NPC has a dedicated head visual.
Additionally, the head node is separately checked during refined
collision detection, because it is not considered in the model trace ray
(at least in Gothic 1).
This whole approach might impact performance, because the bounding box of
all models is calculated each frame.