Reduce auto created ammo items for NPCs
If an NPC is created with a ranged weapon but without matching muntion, 50
munition items are created automatically (in oCNpc::Refresh). This
introduces an exploid when using the re-usable projectiles feature,
because ransacking downed/killed NPCs offers too many projectiles. This
commit reduces the number of auto created projectiles to five. Since
munition is created on the fly with B_SelectWeapon() this does not have
much impact. Even less so, because most NPCs are created with a specific
number of projectiles in their inventory defined in their instance
scripts, thus not being affected at all.