Set stricter targeting system for spells
So far, focus collection for free aiming spells was done by trace ray with
bounding box detection. This makes it harder to focus specific NPCs
standing in a group. Now, the focus collection is done as it is for ranged
combat: by NPC model. Whether this makes targeting spells too difficult
has to be tested with this commit and reverted in case this mechanism is
too strict.
Additionally, the node targeting of spells has been removed for free
aiming spells (e.g. a spell flies towards Bip01 always) to prevent rare
behaviors in which a spell does not fly into the direction as intended.
This latter change is also subject to testing, because it removes the
focus and enemy properties of the player for the moment of casting to
leave the target in the spell instance blank. Whether this causes problems
(e.g. spell can only connect to NPCs that are defined as the spell target)
has to be tested. Nevertheless, spells that support free aiming do not
require a focus to begin with, that's why this change should cause no
harm.