Add free aiming and free movement for focus spells
Now spells that require a focus to be cast (targetCollectAlgo ==
TARGET_COLECT_FOCUS) also support free aiming and free movement. Nothing
changes in their usage except that their focus collection is more strict and
a reticle is shown (as well as the ability to move while aiming/casting).

Additionally, the previous commit is partially reverted. The attempt to
disable node-specific trajectories (which was faulty and did not work) was
removed again.