Make free aiming hooks dynamic, add emergency lock
When disabling free aiming in the game menu, all traces of it will now be
removed. That includes engine hooks, which will be re-initialized when
activating free aiming again in the menu. This was done to allow
completely disabling free aiming, in case of compatibility issues or bugs
related to free aiming in an already released mod. This provides a
fallback until a patch for the mod-project is released.
Since this only concerns the free aiming feature, but not any of the other
possibly initialized sub-features (re-usable projectiles, critical hits
and extended collision behaviors), an "emergency lock" can be activated
from the Gothic.INI file to block the initialization of any GFA
feature completely.