Re-write critical hit feature
Critical hits no longer require to define a single "weak spot" (model
bone) that will be checked for intersection of the projectile trajectory,
but instead already provides the model node (bone) that was hit by the
projectile. This way the config funtion can react to multiple different
bones.
This allows the hit detection to go way beyond simple critical hits and
also removes the need for a separate critical hit event funtion. Now the
user has a lot more freedom in the critical hit configuration.