Fix #191 fix crossbow aiming with low FPS
The source of the issue was not the aiming condition, but the falling
detection in the function GFA_IsActive. The player model distance from the
ground was checked too strictly. The crouching animation of the crossbow
T2 stance introduces a higher distance for a short period of time which
was caught more often with lower FPS.