Temporarily provide colliding projectile
The collision configuration functions lacked access to the current
projectile and the position of collision. Both can now be accessed via
the pointer of the colliding projectile which is temporarily provided
with the global variable GFA_ProjectilePtr. It can be accessed with

    var oCItem proj; proj = _^(GFA_ProjectilePtr);
    Print(proj.name);

This unconventional method of providing information to the config
functions preserves the function signatures and thus prevents issues in
backward compatibility.